Added yt darkmode and made skull emote link to homepage (#17)
* Added random quotes, 20 of em!

Added random quotes!

* added anthony's changes

and fixed quotes lmao they were so broken

* Update quotes.js

Co-authored-by: Anthony Morris <anthony@amorrissound.com>

* Update popup.html

Co-authored-by: Anthony Morris <anthony@amorrissound.com>

* Update popup.html

Co-authored-by: Anthony Morris <anthony@amorrissound.com>

* Added youtube darkmode, and made logo link to homepage

Added youtube darkmode support, though could probably be improved in the future. Does not break anything.

Also made the skull emote a link to the homepage of the site you are on. Because it was not working for youtube before and it bugged me.

* Update death-clock.js

Co-authored-by: Anthony Morris <anthony@amorrissound.com>

* Update death-clock.css

Co-authored-by: Anthony Morris <anthony@amorrissound.com>

* Update death-clock.js

Co-authored-by: Anthony Morris <anthony@amorrissound.com>

* More elegant ala anthony

Changed darkmode to add a css style dynamically. Very neat!

* anthony changes

changes tony made

* Add proper white space

Co-authored-by: Anthony Morris <anthony@amorrissound.com>