feat: re-write app with Plasmo (#22)
### tl;dr

- [x] Re-writes the entire extension using [Plasmo](https://plasmo.com)
- [x] Re-designs entire extension

### Description

Social Media Death Clock was tedious to maintain. Replacing logos on social media websites ended up not being worth the effort. A solution to this was coming up with an entirely new design. For v2, we do that! 🤘🏻 

This design takes a little more of a brutalist approach. I think it's fun and it also draws your attention in when you go to a social media site. 😅 

Once https://github.com/PlasmoHQ/plasmo/issues/21 is fixed, I'll test the production build and then deploy it to the appropriate web stores!